[#150] refactor: 불필요한 infra 폴더 삭제
