+++ collector/src/main/java/com/wherecar/collector/gpslog/application/GpsLogServiceImpl.java
@@ -4,8 +4,8 @@ import com.wherecar.collector.car.domain.Car;
 import com.wherecar.collector.gpslog.application.dto.GpsLogRequest;
 import com.wherecar.collector.gpslog.domain.GpsLog;
 import com.wherecar.collector.gpslog.domain.GpsLogFactory;
-import com.wherecar.collector.car.infrastructure.infra.CarReader;
-import com.wherecar.collector.gpslog.infrastructure.infra.GpsLogStore;
+import com.wherecar.collector.car.infrastructure.CarReader;
+import com.wherecar.collector.gpslog.infrastructure.GpsLogStore;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
