+++ device_db.json
@@ -1,1 +1,1 @@
-[{"mdn":"01234567890","vrp":"12가3456","totalDistance":"0"}]
+[{"mdn":"01234567890","vrp":"12가3456","totalDistance":"1385"}]
